+++ src/main/java/proxima/informatica/academy/seventh/service/CandidateStatesService.java
@@ -48,7 +48,7 @@ public class CandidateStatesService {
 		return response;
 	}
 	
-	public CandidateStates selectById(int id) {
+	public CandidateStates selectById(long id) {
 		CandidateStates candidateRetrived = new CandidateStates();
 		candidateRetrived = (CandidateStates)candidateStatesRepository.findById(id);
 
@@ -77,7 +77,7 @@ public class CandidateStatesService {
 		return candidateStatesRepository.findAll();
 	}
 	
-	public boolean deleteById(int id) {
+	public boolean deleteById(Long id) {
 		return candidateStatesRepository.delete(id) ;
 	}
